test(libs): upgrade jest to v27
Mostly this was a non-issue and the upgrade required no changes. However, in `utils`, we are using the `jsdom` test environment because a lot of `utils` expects browser things like `File` and `FormData`. In jest v27 they removed `setImmediate` and `clearImmediate` from the globals since they're only really supported by IE10/11. Since one of the packages we depend on uses them, we had to add them back.

They also seem to have removed the "`done` callback" style of async tests, so this fixes the places we were using that to await a promise.