build: depend on `pnpm`
This will make it easier to ensure we're using the same version of `pnpm`, though we do have a bootstrapping problem. In CI we use pnpm@6 to install pnpm@5 globally. Eventually we'll want to use this more specific locked-down version, but for now this just plants a flag in the ground and allows tooling like this to use the right version within the monorepo: https://github.com/eventualbuddha/dotfiles/pull/4.