fix(hardware): account for devices with no battery
In production all our devices have batteries, but this may not always be true. In addition, running in development on a desktop means there will be no batteries. Frontends will now check that there is a battery before trying to check its status.

There is a parallel change in kiosk-browser to support this.