fix(em): configuration race condition
`saveElection` in `AppRoot` calls `setElectionDefinition`, which triggers an immediate re-render. In this render, `electionDefinition` will be set but `configuredAt` will not. This is a problem if the render path tries to use `configuredAt`. To fix this, we render `UnconfiguredScreen` when either `electionDefinition` or `configuredAt` are unset.

Additionally, this replaces the empty string placeholder value—which is not a valid ISO8601 timestamp—with `undefined` and updates asserts to account for that.