feat(cdf): build based on both XSD and JSON schemas
The problem with using only the XSD schema is that it doesn't properly encode entity references, i.e. one "complex type" cannot reference any other type by name. It uses a generic IDREF type that doesn't actually say what type it references. This may be fine for navigating a concrete XML document, but is bad for static generation of code. The JSON schema does actually contain a reference to the type definition, but not the documentation. So this commit combines the two favoring the JSON schema since that's what we're actually going to be using in practice, and just uses the XSD one for documentation.