fix(ui): cancel reporting URL update on unmount
In the event that `PrecinctScannerTallyQrCode` is unmounted after the call to `window.kiosk.sign` is made but before it resolves, the call to `setResultsReportingUrl` after it will cause this error:

> Can't perform a React state update on an unmounted component

This change fixes that by ensuring that the `sign` call does not resolve or reject (i.e. is canceled) if the component is unmounted.