refactor: rename `hmpb-interpreter`
We're likely to have different ballot interpreters for different types of ballots. Until now we've just had our own automatically-generated ballot. We're now building support for ballots with timing marks, which is completely different from the approach we've taken with our own ballots. Therefore I believe it makes sense to create different libraries to support different ballots when the differences between them are large, as with ours vs. timing-mark ballots.

- Rename `libs/hmpb-interpreter` to `libs/ballot-interpreter-vx`.
- Update all references (just a global search & replace).
- Re-order some sorted lists.

Refs #1343