refactor(utils): rename ballot package zip function
In #1413 I changed the argument type from `ZipFile` to `Buffer` but didn't change the name. To clarify what the argument is I changed it to mirror the `FromFile` variant as `FromBuffer`.

I also added a test for this function specifically and removed the intermediate `ballotPackageUtils` export within the `/utils` package, but preserved it for external consumers.