refactor(utils): export ballot package fns directly
It seems we were using the intermediate `ballotPackageUtils` as a way to fascilitate mocking. Instead, this changes it to just export the functions directly and mock just that one function in the test that needs it. The mocking is ugly because it should be: this isn't a use case we'd like to see proliferate.