refactor(scan): extract `DbClient`
`Store` has been responsible for both the business logic and the database access jobs. This commit extracts the database access job into a new class, `DbClient`, and renames its functions to be less namespaced. For now I retained most of the `db*` methods in `Store` to keep the changes to existing files isolated to `store.ts`. At a future time we should remove them and provide more semantic access to the database to any callers using them. This commit also creates new tests for `DbClient`.