feat(types): add validation for explicit write-ins
Such candidates were already technically allowed, but had and still have undefined behavior in the rest of the system. Now they are only allowed if there is a write-in candidate placeholder for each seat in a contest. Future commits will update the rest of the system to account for their optional presence. Eventually we may choose to make them mandatory.