refactor: improve compressed tally types
The tests for the compressed tallies were very hard to read because of the opaque lists of numbers, and to a lesser extent the nonspecific type `number[]` used to represent them.

To improve this I added comments to indicate what each number is for. And the types are now numeric tuples with labeled entries, making it easier to see what field corresponds to what.