refactor(scan): separate template interpretation
I plan to make it so the scan service does not need PDFs in order to do its job. Instead, it will have layout information that is generated by interpreting the rendered PDF as we do now, only that will happen in election-manager. That layout information may also have oval images to compare against–I haven't figured that part out yet.

For now, this makes `Interpreter#addTemplate` require a `BallotPageLayoutWithImageData` object rather than allowing an `ImageData` to be passed. It still has an `interpretTemplate` method to find the layout from an image, but it forwards to a standalone function that we'll be able to use from election-manager in the future.

This commit also splits up the takes-forever-to-run `interpreter.test.ts` file so that the parts can run in parallel.