refactor(scan): use canvas `createImageData`
I originally made my own version of it because we were using `sharp` to do some of the image processing. We're no longer doing that, so we don't need this anymore. In fact, the results of this function are not compatible with `canvas`, so it's better to just generate them using `canvas` directly.