chore(scan): use `esbuild-runner` for `vx` CLI
This CLI only worked when the project was built and required a rebuild for changes to the `.ts` files to be reflected. This commit switches to translating to `.js` on the fly using `esbuild-runner`. This would have been a simple change that only touched `bin/ballot-interpreter-vx.js`, but `esbuild` works a little bit differently from `tsc`. In particular, it translates ES imports/exports to CommonJS in a way that makes circular imports not work in some cases where they would with `tsc`. Since the `help` command was referring to `index` and vice versa, this code failed to run correctly under `esbuild-runner`. To fix it I changed the structure to no longer have cycles. While I was in there I improved the tests a bit and stopped throwing exceptions for non-exceptional things like invalid command arguments.