refactor(scan): detect QR codes in `ballot-interpreter-vx`
Refs https://github.com/votingworks/vxsuite/issues/1512

The way we look for QR codes is pretty specific to the `vx` ballot, and doesn't really generalize. At that, QR codes on non-BMD ballots is already pretty edge-case. Because of this it makes more sense for this code to live in `ballot-interpreter-vx`.

Doing this required moving some tests and fixtures, as well as removing the QR code libraries as direct dependencies of `services/scan`. For now we're still calling `detectQrCode` from `services/scan` since we're using worker processes to look for both front/back QR codes at once. The next step will be to move this optimization into `ballot-interpreter-vx` too.