feat(precinct-scanner): Scanner Not Detected screen
When the precinct scanner is disconnected, show a screen to the user and
don't poll for scanner status. Hopefully this will be a useful reminder
and also prevent making hanging requests to the scan service (which is
non-responsive when the scanner is disconnected - though likely we'll
want to fix that as well).