feat(utils): make memory battery status more realistic
When polling for battery status in kiosk-browser, the real value returned by `readBatteryStatus` is always a new object. This currently causes `useDevices` to trigger a re-render of its calling component whenever battery status is updated, which is currently every 3s.

However, when runnning in Chrome we're using `MemoryHardware`, a fake hardware interface. Its `readBatteryStatus` method was returning the same object every time unless the battery level changed. This does not mimick the real one, and therefore was a difference in behavior between Chrome and kiosk-browser that can make bugs appear in one and not the other.

This commit fixes the issue by ensuring that the object returned by `MemoryHardware#readBatteryStatus` is always different, just like with the real one.