fix(scan): handle intentionally blank pages better
If a ballot sheet has contests on only one side then it will have an intentionally blank back page. This page isn't actually blank, but has a sentence explaining that it is being left blank intentionally. It also has a footer with the usual election information and QR code.

Additionally, we have some code that attempts to determine whether a page is blank (in the sense of an empty white page) before trying to look for a QR code. This is partly an optimization and partly to make sure nothing weird ends up on the other side of a BMD ballot.

Putting these together, though, causes a problem. The "blank page" detection works by counting the foreground (dark) vs background (light) pixels and, if the percentage of foreground pixels is less than 0.7% then we consider it truly blank and do not bother looking for a QR code. The threshold of 0.7% is very close to the value for an average "blank" sheet back page of letter size. For legal size, the foreground percent can be even lower.

This hasn't really been a problem because we infer "missing" QR codes when they aren't read properly. However, we can do better. This commit looks in the areas that we would look for QR codes and determines if any of them looks dark enough. If they are dark enough, we treat the page as non-blank and proceed to look for a QR code.