feat(em): export HMPB layout with ballot package
Refs #1558

Pulls `ballot-interpreter-vx` into `election-manager` in order to analyze the layout of the ballots during export. This required getting PDF.js working in kiosk-browser, which required serving `pdf.worker.js` in a way that PDF.js could load it. This required modifying the proxy for `election-manager` to support loading from `pdfjs-dist`'s directory. Otherwise, it was pretty straightforward. In a future commit, I'll modify the `scan` service to accept the layout information from the ballot package.