fix(scan): Fix empty election hash in interpreted BMD ballot (#1560)
* fix(scan): Fix empty election hash in interpreted BMD ballot

This bug surfaced when scanning a live ballot in test mode, since the
adjudication screen tries to parse the interpreted ballot metadata and
the metadata schema expects a non-empty election hash. Since we have
that data available in the interpreter, we can just return it. It seems
like just a weird legacy behavior to return an empty string, as that
code was written a year and a half ago.

Also renames the `HexString` type to `ElectionHash`, since hex strings
could potentially be empty, but election hashes shouldn't be.

* Type update

* fix test