fix(all): remove placeholder empty election hashes
We had these all over the place, primarily in tests. We also had them when reading the old-style URL-based QR codes on HMPBs, because they don't contain a partial election hash. So this also changes that code to expect an election definition to return the hash properly. That code will hopefully be removed soon anyway.