test(em): prepare for ballot package updates
In trying to get the tests working for #1585 I discovered that the way `app.test.tsx` was testing the ballot package export was broken, but only if there were a series of files being added to the package that were non-empty. `zip-stream` appears to operate synchronously when the files added to the zip stream are empty, but asynchronously when there is data to actually compress. Because we'd mocked `printToPDF` to resolve to an empty buffer, the file data has so far been empty. Writing the layout information for each ballot was not empty, and so we triggered the asynchronous behavior of `zip-stream`. This meant that when we tried to check that the download had completed by looking at the calls to `kiosk.log` we would not find it.

I tried a variety of things to get it to wait long enough, but in the end the only thing that worked was to call `jest.advanceTimersByTime(0)` _while fake timers were disabled_. This, of course, made jest issue a warning but it did work. My theory as to why is that when the application is started and interacted with using fake timers, that the `zip-stream` library was caching a reference to the fake `setTimeout` somehow and that, despite calling `jest.useRealTimers` later, it still held on to the fake timer which would still respond to `jest.advanceTimersByTime`. I didn't dig deep enough to prove this, but decided to just cut my losses and remove the ballot package testing from `app.test.tsx` since it's already pretty well tested by its component test. This removes the `jest.useRealTimers` call that was mucking things up, too. In general we should _not_ mix real and fake timers in a single test or, even better, in a single test file.