chore(deps): Add fsevents as optionalDependency to fix prod build
There was a prod build error that had been ignored for a long time. fsevents is an optional dependency of some vxsuite subdependencies. fsevents is only used on Mac, not Linux, so it should have been skipped, but for some reason the prod build script was trying to install it. This explicitly names it as a top-level optional dependency, which seems to make it get skipped correctly.