feat(scan): add library for NH Accuvote ballots
This does not integrate anything into VxAdmin or VxScan yet. The scanning part of this is not New Hampshire-specific, but generalized to Accuvote ballots as far as I've been able to test. Nevertheless, I'm putting it into a `nh`-suffixed library because we don't have any other need to scan Accuvote ballots yet. If we do, it can be separated from the XML conversion stuff (which may _also_ not be NH-specific, but I don't know).

The general idea for data modeling is that it introduces a new part of an election definition for grid-based ballots that defines the grid size and where each contest/option exists in the grid.

Scanning Accuvote ballots happens like so:
1. find the timing marks to determine the grid size and shape
2. match against an empty bubble image at each possible contest/option location to determine a score
3. interpret the marks based on the mark thresholds

The matching algorithm in step 2 is similar to the one from `ballot-interpreter-vx`, but likely better at actually lining up the scan with the template. Because of this we can get really low scores for non-filled in bubbles, which allows us to have pretty low default thresholds. I've picked 3% marginal and 5% definite based on my testing. In the future I plan to port this matching algorithm to `ballot-interpreter-vx` or extract it to have them both share it.

This was done all at once because I had done a bunch of work on a `nh-demo` branch with a bunch of poorly-separated `WIP` commits. Sorry about that.