fix(bmd): cancel printer status request on unmount
It's possible that the `readPrinterStatus` call will not resolve until after the component has unmounted, in which case the `setPrinterStatus` call after it will trigger a warning about a potential memory leak. To prevent this warning in the tests we had an async check for the printer status in the UI which only appears after the request resolves.

This commit fixes that by using `makeCancelable` to cancel the printer status request on unmount, then unmounts the component at the end of the tests that don't use the printer status. This ensures that the warning is not issued and that the `makeCancelable` is working properly.

See https://github.com/testing-library/react-testing-library/issues/667 for a bit more context on this.