refactor(scan): replace `jsdom` with `xmldom`
It turns out that including `jsdom` in the `election-manager` build was causing a runtime error in
Chrome:

```js
const AsyncIteratorPrototype = Object.getPrototypeOf(Object.getPrototypeOf(async function* () {}).prototype);
```

That line appears in `whatwg-url`, which is included by `jsdom`. For some reason the outer `Object.getPrototypeOf` in that line crashes due to its argument being `null` or `undefined`, only as far as I can tell the argument value is not either of those things.

Since I was using `jsdom` for parsing XML and generating SVG strings, I can replace it with `xmldom` which has a lighter footprint (and zero dependencies).

Additionally, this moves `parseXml` out of `src/utils.ts` and into `src/cli/convert/index.ts`, which is the only place in non-test code that we actually needed XML parsing.

The downside is that the DOM returned by `xmldom` is not as rich as the one returned by `jsdom`, i.e. it doesn't have things like `querySelector`, so code that used such features had to be rewritten.