fix(scan): use correct pluralization
I noticed this while watching the video from https://github.com/votingworks/vxsuite/pull/1646. I opted not to break the sentence up with the condition in the middle because that's problematic for localization tools and I think it makes it harder to read, though at the cost of some duplication.