fix(scan): confirm delete batch on enter
Cause the 'Yes, Delete Batch' button to be focused when showing the modal so that pressing Enter or Space will confirm the delete. Escape already works to dismiss it, but with this change you can also press Tab to focus the Cancel button and select it with Enter or Space.