fix(scan): ensure write-in adjudication works front & back
There was an outstanding bug with write-in adjudication where if both the front and back had write-ins to adjudicate _and_ the front had more contests with write-ins than the back then it would crash. Even if the front had the same or fewer it still wouldn't behave correctly, and would instead start at the index of the last contest on the front when adjudicating the back.

The problem was that we were managing the `selectedContestId` at the wrong level of the component hierarchy. This changes it to manage the selected contest ID at the `WriteInAdjudicationScreen` level so that it can reset it properly when finishing a contest rather than letting a stale value from the previous page's contest list stay in state.