fix(scan): make write-in adjudication work for timing mark ballots
This is only an approximate fix. It generates the expected layout for the timing mark ballots, but does not base it on the real interpreted ballot. Thus, it's possible for the coordinates to be somewhat off, especially if a slightly incorrect scaling factor was used. In the future, we should generate a layout based on the `interpret` output for timing mark balltos, which should include positional information for the ovals at minimum.