fix(types): make `CandidateId` encompass `WriteInId`
As-is `CandidateId` is used in some places where a `WriteInId` is valid, such as in `CandidateContestOption`. For now, I think it makes sense to expand the allowed values for `CandidateId` to include `WriteInId` rather than making a bunch of one-off exceptions. In the future I think we need to improve our data model with regard to write-ins, and we can clean this up more then.