fix(scan): correct write-in adjudication for skipped contests
Fixes a bug where we had an index of the current contest that was being used to index both a list of _just the contests with write-ins_ as well as _the list of all contests on the page_. As long as these were the same we didn't have any problems, but if a contest with a write-in appeared after a contest without then this would fail.

Adds documentation to the props and changes the name of `contestIds` to `allContestIds` to make it clear that the list includes the IDs of all contests on the page, not just those with write-ins to adjudicate.