fix(scan): handle write-ins that may not appear in order
In the New Hampshire ballots we have, the write-ins are not defined top-to-bottom but bottom-to-top. Our code assumed that they would always be top-to-bottom, so in any contest with more than one write-in the adjudication screen would crop the wrong section of the image. This fixes it by encoding the contest/option definition along with the layout, allowing the client to properly match up the write-in to be adjudicated with the correct layout information.