feat(scan): enable scanning right-side targets
Closes https://github.com/votingworks/vxsuite/issues/1505

Changes `findTargets` to accept a `targetMarkPosition` which determines whether it starts looking offset from the left or the right side of the contest box.

Updates the `layout` CLI command to find the targets and annotate them in the output. It also now takes an option to specify whether to look on the right or left for the targets.

Adds a basic test for scanning a filled-in ballot with the right-side targets.