feat(lint): enable `import/no-cycle` rule
Aside from being generally recommended against, it actually breaks `esbuild-runner` if the code it's running has cycles. NodeJS and Webpack handle them fine as long as the mutual dependency is deferred past the initial module load, so it hasn't been a problem yet. But I'd like to move away from `ts-node` and toward `esbuild-runner` because it's much faster, and so we need to get rid of cycles.