refactor(all): replace `crypto` with npm deps
Refs #1314

In several packages we import the `crypto` NodeJS built-in library. This works as expected in NodeJS, but for the browser we rely on CRA via webpack to bundle a polyfill. To reduce our reliance on CRA and webpack, this commit switches to explicit polyfills from npm.

In all but the `lsd` module (because it's NodeJS-only), replaces uses of `crypto` with `randombytes` (which uses `crypto` in NodeJS and `globalThis.crypto` in the browser) or `js-sha256` (which likewise uses `crypto` in NodeJS and its own implementation in the browser).