fix(bmd): Use should('not.exist') instead of should('not.be.visible')
Cypress 6 includes a fix that changed the behavior of their `.visible`
implementation. https://github.com/cypress-io/cypress/issues/205

The impact of that change is that cypress will always throw an error
when trying to assert on an element that does not exist in the DOM,
unless you are asserting should('not.exist').