refactor(bmd): Separate non-interactive font size state
Currently, BMD uses an app state field `userSettings.fontSize` to control the font size for the entire HTML document. This feature exists so that voters can increase/decrease the font size as needed. However, it's also been used to set a large font size for non-voter specific screens where there's no interactive setting of font size. This is causing unpredictable behavior of the font sizes when refactoring app state for auth changes.

The ideal solution would be for these screens to just have styling that sets their font size. However, all of the styles in BMD use `rem`, which means the font size is tied to the HTML document font size. So fixing that will be a larger change.

For now, this PR separates out the logic for interactive font size changes (which should only apply when the voter-facing UI is rendered) from non-interactive font size changes.