lint(async): enable `@typescript-eslint/require-await`
Fixes a bunch of violations, including those that cascaded from removing the `async` on functions making the calls of those functions no longer requiring `await` etc.

I marked a bunch of `async` without `await` as ignored because they were implementing an interface either as a mock or fake version of something. Others were ignored because of async issues in tests that it was easier to ignore for now.