build(res-to-ts): do not build `res-to-ts`
Since it can run without building, set it up as a build-less library. This will work with something like Vite.js, but not with CRA/webpack. That's fine since this library is really just a CLI, but we can't do this with other libraries yet.

This avoids `tsconfig` reference cycles.