ci: validate CircleCI config
It's too easy to accidentally forget to add new packages to the config, and even when you do it's easy to forget to add the job to the workflow. This does some minimal validation of the config to prevent these sorts of issues.

It also found an issue! `cdf-types-cast-vote-records` was not being tested. Oops.