test(all): work around async ordering issues
These tests started failing due to some difference in how async operations happen in tests with `@testing-library/react`. What would happen, as an example from Election Manager, is that the test would click the "Save" button, which would trigger an async operation before a 2s sleep, then `jest.advanceTimersByTime` would be called in the test before the async operation as a result of "Save" could resolve, and only then would the 2s sleep be triggered. However, the call to advance the time by that amount had already happened, so the test just times out looking for the state after that 2s sleep. To fix it, we wait for the state immediately before the 2s sleep which forces the `advanceTimersByTime` call to happen after the 2s sleep has started.

The issues in BMD, BSD, and Precinct Scanner are all similar to the above.