test(ui): fix `setResultsReportingUrl` race condition
It's unlikely to happen, but it's possible that the `useEffect` that signs the VxQR results payload will set the results reporting URL to a stale value if two back-to-back renders have `kiosk.sign` calls that return out of order.

We were wrapping the `kiosk.sign` result `Promise` in `makeCancelable`, but that only prevents calling `setResultsReportingUrl` if the component _is unmounted entirely_. It does not stop an earlier render from clobbering the URL update from a re-render.

The fix is to track whether the current render is stale. `useEffect` will call the returned function both on unmount _and on re-render_. We use that to mark a render as stale and do not update the URL if it is stale by the time the `kiosk.sign` call resolves.

If this is something we find ourselves doing often it may make sense to create a hook or other utility to help consolidate the logic.