fix(eslint-plugin-vx): correct `gts-spread-like-types` false positives
We've been seeing false positives with this rule somtimes. Turns out it's a difference in the TypeScript version from when we upgraded from 4.2.x to 4.6.x, but I didn't catch it during the upgrade because I forgot to upgrade the linter package's copy of TypeScript. This fixes it so it should work in both versions, though we only really care about 4.6.x.