build(bsd): add `vite` as a dev & build option
This adds `vite` as a dev & prod build option alongside the (un-ejected) CRA dev & build setup. This allows us to test that everything is working well before we fully commit to `vite`. I did this one differently from `bas` and `bmd` because we don't _really_ need to eject to add the `vite` option.