fix(scan): increase `pdfToImages` PDF compatibility
This has probably been broken for a while with certain PDFs. Depending on the features used in the PDFs, different code paths are taken. It just so happens that the PDFs we've used up to now have not required the `canvasFactory` implicitly used by `pdfjs` to actually work, presumably because it could just paint directly on the canvas that was given and it didn't need any intermediate working canvases. The PDFs we got from NH actually do have some feature that requires an intermediate canvas, and so it calls to create a new canvas and fails with "Reference Error: document is not defined". `pdfjs` tries to use the DOM to build `canvas` objects by default, which is sensible. But in NodeJS it just doesn't work. Instead, we now provide a factory that uses the `canvas` NPM package to build canvases.

This bug doesn't affect the _other_ `pdfToImages` implementation in Election Manager because there we actually do have the DOM and the default factory works fine.