feat(scan): accept scans of any size and scale them
I ended up doing a fairly large refactoring of our image-handling code in the NH ballot interpreter as I was trying to figure out the best way to scale the incoming scans. It turns out it's best to scale them before converting to an `ImageData`, since with an `Image` you can use `drawImage` with a size that will rescale it.

Additionally, I no longer explicitly convert to grayscale. I just assume that all images are grayscale and use the 4-channel RGBA images in memory, reading the R value for the luminosity of the pixels.

Fixes #1896