feat(types): support multi-party endorsement
Replaces `Candidate#partyId` with `Candidate#partyIds` to allow specifying that a given candidate is endorsed by multiple parties. This involved a _lot_ of search and replace in fixture data, and updated a few places where we were showing the previously single party a candidate was associated with. This doesn't intentially change reporting at all, though I imagine there may be future changes we want to make to support reporting which under which party a candidate was selected on a HMPB.

Fixes #1953