fix(scan): ignore mismatched XML/PDF ballot size
Rather than simply removing the error condition for the mismatched ballot size, I opted to make the conversion process a bit more error-tolerant while still surfacing the errors. Now it'll build the election definition as much as it can while also collecting issues as it sees them, returning the issues along with the (potentially) partial election definition. We consider the mismatched ballot size to be a non-fatal issue, so if that's the only issue then `success` will be `true`. All other issues are, for now, fatal but will no longer block generating an election definition.

This commit also updates the Amherst election definition to be as complete as it can be at this point, which means excluding the constitutional question. But that means ballots should now be scannable, just without the constitutional question.

Fixes #1987