fix(bmd): Don't load tally from pollworker card if already loaded
There was a loop happening because the pollworker screen would load the
tally once, then print it, then load it again before clearing it from
the card (due to re-renders triggering the useEffect hook multiple
times).

To fix, we put in a simple guard to make sure we only load the data if
we haven't already loaded it.