fix(scan-nh): find timing marks more reliably
The old approach of finding timing marks by scoring blocks has a major downside: if the blocks are sufficiently cropped they will fall below the threshold and no longer appear as timing marks. This happens regularly when using the precinct scanner. Reducing the threshold introduces too much noise. Though it's possible we could have inferred the missing ones on one side by looking at the other, there's no guarantee we'll see the ones on that side properly either.

Instead, we now look for the gaps between timing marks and determine the lines of blocks based on that. This is quite reliable on the left and right, but less so on the top and bottom. I haven't done extensive testing yet and I'm sure we'll need to iterate, but this seems both faster and less sensitive to distortions.