build(em): add stub for `fs` module
`browserify-fs` worked fine in development as a replacement for `fs`, but doesn't work in production. This is due to a circular dependency in `readable-stream`, which `browserify-fs` depends on. Since we don't actually want to have a working simulated file system, we can just create a stub. Rather than doing it by hand I wrote a simple script to generate it. I also updated the stub for `glob` using it.