fix(frontends): ensure `path` is loaded correctly
For some reason `import path from 'path'` would get built for production such that `path` was just an empty object. Using named imports makes it break at build time, which is a better outcome. To fix the build, we alias `path` to the one installed from NPM instead of the NodeJS built-in `path` module.

Closes #2045