fix(scan-nh): handle stretched ballot images better
This commit relaxes some of the constraints we had that turned out to be overly aggressive when interpreting images from the precinct scanner.

We require that the top/bottom timing mark lines' angles be within an expected range based on the left/right lines' angles. When the ballot image is stretched this angle can be more off than expected, e.g. 1.3° or so (> 1°, the previous threshold).

Additionally, the cropped side timing marks might cause the scanning of the top timing marks not to consider the gap between the top left corner and the mark to its right as a real gap between timing marks. Thus, the left and top lines of timing marks did not have a shared sense of where the corner was. We're now more careful to ensure that the top corner timing marks contain the intersection of their respective best-fit lines.

Finally, the side timing mark gaps were not being correctly identified in some cases because they were too narrow (<50% the expected width). Now, we keep any left/right gaps for consideration as long as they are wide enough and/or tall enough. We don't do the same when looking for the top/bottom timing mark gaps because it creates too many false positives.