fix(scan-nh): improve handling of folded ballots
It seems the folded ballots we have do not create a visible line in the scanned image, but they may be causing the rate at which the scanner feeds the paper through to vary as it's going. This leads to slight vertical stretching at various parts of the ballot image. In the new fixture images added here, the issue was that the algorithm to place timing marks between gaps ended up placing two timing marks essentially stacked on top of each other. The simplest fix for this is to replace the deduplication method I was using, i.e. "is the center of the new timing mark inside any existing timing marks?", and replace it with one that's more likely to catch duplicates, i.e. "does the new timing mark overlap at all with an existing timing mark?".