fix(nh): handle more ballot distortion issues
Problem 1: Ballot image stretching can be quite pronounced with the precinct scanner, with the left and right being quite straight but the top or bottom being quite slanted. I increased the allowed angle error for the top from 2° off what we expect to 3° as I found one that had a 2.4° error.

Problem 2: Some ballots have timing marks whose edges are a little fuzzy. This can lead to the gap lines not being properly grouped, which causes a skinny timing mark gap to appear alongside the main real gap. The best fit search algorithm previously tried to maximize the number of rects it passed through as long as the error was not too great in doing so. Sometimes this would lead to the best fit line skewing to the edge of the timing marks to try to include these skinny extra gaps, since doing so means the best fit line would pass through both the real gap (albeit imperfectly) and the skinny one (quite accurately). I changed the scoring mechanism of the best fit line algorithm to scale the score proportional to the dimensions of the gap rectangle, creating a much greater incentive to accurately fit to the real larger gap rectangles.

Problem 3: Sometimes the ballot stretching would appear in between timing marks in such a way that we would infer an extra timing mark where there should not really be one. To fix this, we still infer missing timing marks, but then remove the ones that appear not to be real (i.e. not very filled in).