fix(nh): properly flag undervotes for multi-party endorsements
Closes #2072

The data model was changed in #1984 to support multi-party endorsements, but we weren't properly handling it in the NH code that turns marks into adjudication reasons. The result was that we'd determine that the contest was undervoted if only the second grid position was filled in for a given option. The votes were recorded correctly, but an error would be shown.

The fix is to examine all the grid positions that map to the contest/option and use the most marked one to determine the mark status.