fix(scan): show overvote/blank sheet as overvote
Since each side of the sheet has adjudication reasons generated separately they can have a `BlankBallot` adjudication reason even though the ballot as a whole may not be blank. To work around this we process the whole sheet together when building the scanner result. This really should be done on the backend, but this is a sufficient temporary fix.

Closes #1638