fix(scan): detect BMD ballots scanned as 14/17"
We currently always print BMD ballots on letter-size paper. We scan centrally using a fixed size according to the election definition. Thus, when scanning a BMD ballot in an election configured with non-letter paper we will have images that have extra junk above or below the actual ballot. To account for this, we now check where the QR code could be when printed on letter but scanned as non-letter, but only when scanning non-letter size paper.

Fixes #911